Improve Makefiles for "make hpct" Make cast to double explicit in gputest.c Set up install files for rice's ufront machine

+++ TEST/src/gputest.cc
@@ -127,7 +127,7 @@ void
 init(double *p, size_t size)
 {
   for (size_t i = 0; i < size; ++i) {
-    p[i] = i+1;
+    p[i] = (double) (i+1);
   }
 }
